+++ ext/mbstring/libmbfl/mbfl/mbfl_string.h
@@ -31,6 +31,7 @@
 #ifndef MBFL_STRING_H
 #define MBFL_STRING_H
 
+#include <stddef.h>
 #include "mbfl_defs.h"
 #include "mbfl_encoding.h"
 #include "mbfl_language.h"
@@ -42,7 +43,7 @@ typedef struct _mbfl_string {
 	enum mbfl_no_language no_language;
 	enum mbfl_no_encoding no_encoding;
 	unsigned char *val;
-	unsigned int len;
+	size_t len;
 } mbfl_string;
 
 MBFLAPI extern void mbfl_string_init(mbfl_string *string);
